Get rid of optimization module
Move all of the code into FlowSolver
